There is A Lot So Read Description
Changed Create.js recipe to use the new separated macros.
Added the new testing for recipe create along with its documentation in the yaml file.

Created the create/tag api, its testing and inclusion in the yaml file.

Created search/tag and its unit testing. There is no documentation for it yet.

Changed get/recipe to not return the id.

Search.test.js for recipe is not correct as of right now. Needs to be touched.

Played around with adding images and how it works. Unsure if it works as intended.