Check description.
Log-in adjusted to now use JWT. After successfully entering the right credentials, the api will generate a JWT and return that to the front end. Later fixes will ensure that the backend verifies it on later calls.

Changed search recipe to only include necessary parameters. There is commented code to change the image back from its buffer form so I will test that later.
Search also includes the firstidx and lastidx.

Unit testing has been changed for search but I cannot say it is finished.

Created update recipe and its unit testing. Changes may occur so it is not in testing for now.

Included jsonwebtoken in the packages.

Create recipe changes were simply to better see a bug occurring when trying log in.