Description
Changed update recipe to function correctly. id is the only necessary input and it checks for other inputs and type checking if necessary. Tested it using Postman to see if it updates.

Added update recipe to server.js.

Played around with some image code in get.js. Cannot test to see if it works so it is commented out as of right now. Other image processing method is still there.

Unit testing is not done for either search or update as they must be changed.