Description.
For Search Recipe:
Fixed using only calories for search query.
Changed tagId to use hasEvery instead of equals for filtering.
Fixed the macro checks to be more concise and structured.
When there are no tagIds from the frontend, it runs. That is to say, if the user does not create tags, don't send a tagId item to the API.